refactor: add login folder and move related files in it
