+++ androidApp/src/main/java/com/example/tasky/android/agenda/viewmodel/AgendaDetailsViewModel.kt
@@ -20,7 +20,7 @@ import com.example.tasky.android.agenda.screen.AgendaDetailsScreenEvent
 import com.example.tasky.android.agenda.screen.AgendaDetailsScreenState
 import com.example.tasky.android.agenda.screen.AgendaDetailsScreenType
 import com.example.tasky.android.utils.IImageCompressor
-import com.example.tasky.manager.SessionManager
+import com.example.tasky.login.domain.SessionManager
 import com.example.tasky.model.onSuccess
 import com.example.tasky.util.toLocalDateTime
 import kotlinx.coroutines.channels.Channel
