+++ androidApp/src/main/java/com/example/tasky/android/login/viewmodel/RegisterViewModel.kt
@@ -4,17 +4,19 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasky.android.login.screen.RegisterScreenEvent
 import com.example.tasky.android.login.screen.RegisterScreenState
-import com.example.tasky.model.login.RegisterBody
+import com.example.tasky.login.data.model.RegisterBody
+import com.example.tasky.login.domain.ILoginRepository
 import com.example.tasky.model.onError
 import com.example.tasky.model.onSuccess
-import com.example.tasky.repository.ILoginRepository
 import com.example.tasky.util.Validator
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-class RegisterViewModel(private val loginRepository: ILoginRepository) : ViewModel() {
+class RegisterViewModel(
+    private val loginRepository: ILoginRepository,
+) : ViewModel() {
     private val _screenStateFlow = MutableStateFlow(RegisterScreenState())
     val screenStateFlow = _screenStateFlow.asStateFlow()
