+++ shared/src/commonMain/kotlin/com/example/tasky/login/data/model/Login.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.model.login
+package com.example.tasky.login.data.model
 
 import io.ktor.resources.Resource
 import kotlinx.serialization.Serializable
