+++ shared/src/commonMain/kotlin/com/example/tasky/login/domain/LoginManager.kt
@@ -1,10 +1,10 @@
-package com.example.tasky.manager
+package com.example.tasky.login.domain
 
 import com.example.tasky.dataStore.SettingsKey
+import com.example.tasky.login.data.model.LoginBody
+import com.example.tasky.manager.HttpManager
 import com.example.tasky.model.ResultWrapper
-import com.example.tasky.model.login.LoginBody
 import com.example.tasky.model.onSuccess
-import com.example.tasky.repository.ILoginRepository
 import com.russhwolf.settings.ExperimentalSettingsApi
 import com.russhwolf.settings.coroutines.FlowSettings
 import kotlinx.coroutines.ExperimentalCoroutinesApi
